requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.8.5
django-debug-toolbar==3.2.2
flake8==4.0.1
stripe==2.63.0
drf-yasg==1.20.0